backend/app/models/company.py
from tortoise import fields, models

class Company(models.Model):
    id = fields.IntField(pk=True)
    name = fields.CharField(max_length=255, unique=True)
    registration_number = fields.CharField(max_length=100, unique=True, null=True)
    country = fields.CharField(max_length=100)
    created_at = fields.DatetimeField(auto_now_add=True)
    updated_at = fields.DatetimeField(auto_now=True)

    class Meta:
        table = "company"
