backend/requirements.txt
fastapi
tortoise-orm
uvicorn
pydantic_settings
aiomysql
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]